fix: store timestamps as integer microseconds instead of float seconds

+++ tests/test_utils.py
@@ -127,5 +127,5 @@ def test_build_usage_record_includes_provider_metadata(utils_module):
     assert record["ttft_ms"] == 11
     assert record["status"] == 201
     assert record["total_tokens"] == 15
-    assert isinstance(record["ts"], float)
+    assert isinstance(record["ts"], int)
     assert record["ts"] > 0
